+++ examples/W_energy_est.jl
@@ -1,3 +1,4 @@
+
 using EquivariantModels, Lux, StaticArrays, Random, LinearAlgebra, Zygote 
 using Polynomials4ML: LinearLayer, RYlmBasis, lux 
 using EquivariantModels: degord2spec, specnlm2spec1p, xx2AA
@@ -45,96 +46,6 @@ model = Chain(
 ps, st = Lux.setup(rng, model)
 out, st = model(X, ps, st)
 
-module Pot 
-   import JuLIP, Zygote, StaticArrays
-   import JuLIP: cutoff, Atoms 
-   import ACEbase: evaluate!, evaluate_d!
-   import StaticArrays: SVector, SMatrix
-
-   import ChainRulesCore
-   import ChainRulesCore: rrule, ignore_derivatives
-
-   import Optimisers: destructure
-
-   struct LuxCalc <: JuLIP.SitePotential 
-      luxmodel
-      ps
-      st 
-      rcut::Float64
-      restructure
-   end
-
-   function LuxCalc(luxmodel, ps, st, rcut) 
-      pvec, rest = destructure(ps)
-      return LuxCalc(luxmodel, ps, st, rcut, rest)
-   end
-
-   cutoff(calc::LuxCalc) = calc.rcut
-
-   function evaluate!(tmp, calc::LuxCalc, Rs, Zs, z0)
-      E, st = calc.luxmodel(Rs, calc.ps, calc.st)
-      return E[1]
-   end
-
-   function evaluate_d!(dEs, tmpd, calc::LuxCalc, Rs, Zs, z0)
-      g = Zygote.gradient(X -> calc.luxmodel(X, calc.ps, calc.st)[1], Rs)[1]
-      @assert length(g) == length(Rs) <= length(dEs)
-      dEs[1:length(g)] .= g 
-      return dEs 
-   end
-
-   # ----- parameter estimation stuff 
-
-   
-   function lux_energy(at::Atoms, calc::LuxCalc, ps::NamedTuple, st::NamedTuple)
-      nlist = ignore_derivatives() do 
-         JuLIP.neighbourlist(at, calc.rcut)
-      end
-      return sum( i -> begin
-            Js, Rs, Zs = ignore_derivatives() do 
-               JuLIP.Potentials.neigsz(nlist, at, i)
-            end
-            Ei, st = calc.luxmodel(Rs, ps, st)
-            Ei[1] 
-         end, 
-         1:length(at)
-         )
-   end
-
-
-   function lux_efv(at::Atoms, calc::LuxCalc, ps::NamedTuple, st::NamedTuple)
-      nlist = ignore_derivatives() do 
-         JuLIP.neighbourlist(at, calc.rcut)
-      end
-      E = 0.0 
-      F = zeros(SVector{3, Float64}, length(at))
-      V = zero(SMatrix{3, 3, Float64}) 
-      for i = 1:length(at) 
-         Js, Rs, Zs = ignore_derivatives() do 
-            JuLIP.Potentials.neigsz(nlist, at, i)
-         end
-         comp = Zygote.withgradient(_X -> calc.luxmodel(_X, ps, st)[1], Rs)
-         Ei = comp.val 
-         ∇Ei = comp.grad[1] 
-         # energy 
-         E += Ei 
-
-         # Forces 
-         for j = 1:length(Rs) 
-            F[Js[j]] -= ∇Ei[j] 
-            F[i] += ∇Ei[j] 
-         end
-
-         # Virial 
-         if length(Rs) > 0 
-            V -= sum(∇Eij * Rij' for (∇Eij, Rij) in zip(∇Ei, Rs))
-         end
-      end
-      
-      return E, F, V 
-   end
-
-end 
 
 ps.dot.W[:] .= 0.01 * randn(length(ps.dot.W)) 
 calc = Pot.LuxCalc(model, ps, st, rcut)
@@ -212,4 +123,4 @@ ylabel("ACE energy")
 axis("square")
 xlim([-142.3, -141.5])
 ylim([-142.3, -141.5])
-PyPlot.savefig("W_energy_fitting.png")
+PyPlot.savefig("/figure/W_energy_fitting.png")
