+++ examples/backup/energy_with_multi_species.jl
