+++ examples/data/NiAl_data.xyz
