+++ examples/figure/W_energy_fitting.png
