+++ examples/figure/linear_W_energy_fitting.png
