+++ src/ACEluxpots.jl
@@ -2,4 +2,10 @@ module ACEluxpots
 
 # Write your package code here.
 
+include("staticprod.jl")
+
+# will merge these two into one file
+include("calculator_single.jl")
+include("calculator_multi.jl")
+
 end
