+++ src/staticprod.jl
